requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==24.2.0
bech32==1.2.0
blinker==1.9.0
blockfrost-python==0.6.0
cachetools==5.5.0
cardano-tools==2.1.0
cbor2==5.6.5
certifi==2024.8.30
certvalidator==0.11.1
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
cose==0.9.dev8
cryptography==43.0.3
decorator==5.1.1
docker==7.1.0
ecdsa==0.19.0
ECPy==1.2.5
executing==2.1.0
fastapi==0.115.5
Flask==3.1.0
Flask-Cors==5.0.0
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mnemonic==0.21
mypy==1.13.0
mypy-extensions==1.0.0
ogmios==1.2.1
orjson==3.10.11
oscrypto==1.3.0
parso==0.8.4
pexpect==4.9.0
pprintpp==0.4.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycardano==0.12.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
ruff==0.7.3
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
traitlets==5.14.3
typeguard==4.4.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
